+++ CHANGES.md
@@ -1,12 +1,20 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## 0.2.1 - 2017-12-22
+
+## 0.2.1 - 2018-01-03
+### Added
+- possibility to pass a list of categories to be analyzed (p: --category)
+- identification of singletones (e.g. those features outside any cluster of the category to which the feature belongs; p: --singletons)
+
+### Changed
+- ClusterScan nomenclature regarding the categorical annotation.
+
+## 0.2.0 - 2017-12-22
 ### Added
 - analysis duration benchmark in standard output
 - summmary of the analysis in standard output
-- possibility to pass a list of categories to be analyzed (p: --category)
-- computation of singletones (e.g. those features outside any cluster of the category to which the feature belongs)
+- "CHANGES.md" and "README.md".
 
 ### Changed
 - improve _clustermean_ speed